Fix networkx.Graph to igraph.Graph conversion
When the passed graph has isolates (i.e., nodes which are not adjacent to any other nodes), the conversion fails with the error "networkx.exception.AmbiguousSolution: Disconnected graph: Ambiguous solution for bipartite sets." on the call to `bipartite.sets`. This adds another check to ensure that there are no isolates present.